# module/services/accounts/openapi/api.swagger.json
{
  "consumes": [
    "application/json"
  ],
  "definitions": {
...
    "protobufAny": {
      "additionalProperties": {},
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "rpcStatus": {
      "properties": {
        "code": {
          "format": "int32",
          "type": "integer"
        },
        "details": {
          "items": {
            "$ref": "#/definitions/protobufAny",
            "type": "object"
          },
          "type": "array"
        },
        "message": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "v1FollowResourceRequest": {
      "description": "FollowResourceRequest deliberately carries no subject field. The follower is\nalways the authenticated caller, so the RPC can never be turned into an oracle\nabout another principal — the same reason ListMyAccessibleScopesRequest omits\none. A target the caller cannot currently see is reported as missing, never as\ndenied, so Follow cannot be used to probe for resources that exist elsewhere.",
      "properties": {
        "resourceId": {
          "type": "string"
        },
        "resourceType": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "v1FollowResourceResponse": {
      "properties": {
        "id": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "v1UnfollowResourceRequest": {
      "properties": {
        "resourceId": {
          "type": "string"
        },
        "resourceType": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "info": {
    "title": "saas/policy/v1/options.proto",
    "version": "version not set"
  },
  "paths": {
...
    "/v1/resource-follows": {
      "post": {
        "operationId": "ResourceFollowService_Follow",
        "parameters": [
          {
            "description": "FollowResourceRequest deliberately carries no subject field. The follower is\nalways the authenticated caller, so the RPC can never be turned into an oracle\nabout another principal — the same reason ListMyAccessibleScopesRequest omits\none. A target the caller cannot currently see is reported as missing, never as\ndenied, so Follow cannot be used to probe for resources that exist elsewhere.",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1FollowResourceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1FollowResourceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ResourceFollowService"
        ]
      }
    },
    "/v1/resource-follows:unfollow": {
      "post": {
        "operationId": "ResourceFollowService_Unfollow",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UnfollowResourceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {},
              "type": "object"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ResourceFollowService"
        ]
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
...
}